+++ lib/src/features/profile/profile_page.dart
@@ -216,7 +216,11 @@ class _ProfilePageState extends State<ProfilePage> {
                         padding: const EdgeInsets.symmetric(horizontal: 20),
                         child: Row(
                           children: [
-                            Text(data?.about ?? ''),
+                            Text(
+                              data?.about ?? '',
+                              maxLines: 1,
+                              overflow: TextOverflow.ellipsis,
+                            ),
                           ],
                         ),
                       ),