+++ changelog.md
@@ -1,6 +1,10 @@
+## v1.10.0-beta
+- Animated Media Banner Image
+
 ## v1.9.0-beta
 - Added `Show Score` toggle on setting screen to hide score on Media List Screen
 - Changes to display on Media Search screen in Listview mode
+- Enabled new Splash Screen on Tablet
 
 ## v1.8.1-beta
 - Fixed title sorting